- add initial flare gamefiles - add offset support to BaseLevelObject - add support to load part of image - add support to loading flare animation offsets - add support to position drawables in top/center/bottom of screen - BaseAnimation class - cache images when loading - don't hide cursor when loading new cursor - fix glitches with level view when zooming resizing - fix small bugs - fix small bugs in menu - refactor classes that use animations to use BaseAnimation - refactor Fonts - refactor items - refactor level parser to get map size from map if none provided - refactor PCX decoder - refactor RectTexturePack - refactor SimpleIndexedTexturePack to IndexedTexturePack - refactor texturePack to support setting start index - refactor texturePacks/level/itemClass/playerClass - try to parse bitmap button and then text button (if bitmap button fails to parse) - update gamefiles - update level draw code - update parseLevel to handle invalid data better

+++ src/Game/MapCoord.h
@@ -2,5 +2,5 @@
 
 #include "PairXY.h"
 
-typedef uint16_t Coord;
-typedef PairXY<uint16_t> MapCoord;
+typedef int32_t Coord;
+typedef PairXY<int32_t> MapCoord;
